+++ packages/selector-specificity/test/index.mjs
@@ -1,5 +1,9 @@
-import './example.mjs';
-import './test-compare.mjs';
 import './test.mjs';
+
+import './test-compare.mjs';
+import './test-custom-specificity.mjs';
+
 import './tests-from-bramus-specificity.mjs';
 import './tests-from-keeganstreet-specificity.mjs';
+
+import './example.mjs';