+++ plugins/css-has-pseudo/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 - Updated: Support for Node v18+ (major).
 
+### 6.0.4
+
+_May 11, 2024_
+
+- Updated [`@csstools/selector-specificity`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity) to [`3.1.0`](https://github.com/csstools/postcss-plugins/tree/main/packages/selector-specificity/CHANGELOG.md#310) (minor)
+
 ### 6.0.3
 
 _March 31, 2024_